requirements.txt
streamlit==1.36.0
folium
networkx
osmnx
pandas
shapely
streamlit_extras
streamlit_folium
st-star-rating
google-generativeai